feat: update for 4.30.00.54

+++ source/ti/ble5stack/osal/src/inc/osal_snv.h
@@ -76,8 +76,8 @@ extern "C"
  * TYPEDEFS
  */
 
-  typedef uint8 osalSnvId_t;
-  typedef uint8 osalSnvLen_t;
+  typedef uint16 osalSnvId_t;
+  typedef uint8  osalSnvLen_t;
 
 /*********************************************************************
  * GLOBAL VARIABLES
